BackgroundExecution ported to 64-bits

+++ BackgroundExecution/AppDelegate.cs
@@ -1,6 +1,6 @@
 using System;
-using MonoTouch.UIKit;
-using MonoTouch.Foundation;
+using UIKit;
+using Foundation;
 using System.Threading;
 
 namespace Example_BackgroundExecution
@@ -64,12 +64,12 @@ namespace Example_BackgroundExecution
 			//UIApplication.SharedApplication.SetKeepAliveTimout(600, () => { /* keep alive handler code*/ });
 			
 			// register a long running task, and then start it on a new thread so that this method can return
-			int taskID = UIApplication.SharedApplication.BeginBackgroundTask ( () => {});
+			nint taskID = UIApplication.SharedApplication.BeginBackgroundTask ( () => {});
 			Thread task = new Thread (new ThreadStart ( () => { FinishLongRunningTask(taskID);}));
 			task.Start ();
 		}
 		
-		protected void FinishLongRunningTask (int taskID)
+		protected void FinishLongRunningTask (nint taskID)
 		{
 			Console.WriteLine ("Starting task " + taskID.ToString ());
